+++ tests/testthat/test-interpolate.R
@@ -52,3 +52,10 @@ test_that("can interpolate from a package", {
 
   expect_equal(interpolate_package("test", "bar.md", x = 1), ellmer_prompt("1"))
 })
+
+test_that("errors if the path does not exist", {
+  expect_snapshot(error = TRUE, {
+    interpolate_file("does-not-exist.md", x = 1)
+    interpolate_package("ellmer", "does-not-exist.md", x = 1)
+  })
+})